Description:clean reviewbot
Match-id-385d76e9398df1e32b93c405a4427ea0664de446

+++ common/src/daudio_util.cpp
@@ -97,9 +97,9 @@ std::string GetAnonyString(const std::string &value)
 
 int32_t GetDevTypeByDHId(int32_t dhId)
 {
-    if (dhId & 0x8000000) {
+    if ((uint32_t)dhId & 0x8000000) {
         return AUDIO_DEVICE_TYPE_MIC;
-    } else if (dhId & 0x7ffffff) {
+    } else if ((uint32_t)dhId & 0x7ffffff) {
         return AUDIO_DEVICE_TYPE_SPEAKER;
     }
     return AUDIO_DEVICE_TYPE_UNKNOWN;
