+++ hdf_service/distributed_audio/hdi_service/common/utils/src/daudio_utils.cpp
@@ -112,9 +112,9 @@ int32_t SetAudioParamBool(std::string &params, const std::string &key, const boo
 
 int32_t GetDevTypeByDHId(int32_t dhId)
 {
-    if (dhId & 0x8000000) {
+    if ((uint32_t)dhId & 0x8000000) {
         return AUDIO_DEVICE_TYPE_MIC;
-    } else if (dhId & 0x7ffffff) {
+    } else if ((uint32_t)dhId & 0x7ffffff) {
         return AUDIO_DEVICE_TYPE_SPEAKER;
     }
     return AUDIO_DEVICE_TYPE_UNKNOWN;
