+++ services/audioprocessor/decoder/src/audio_decoder.cpp
@@ -256,7 +256,7 @@ void AudioDecoder::InputDecodeAudioData()
             if (inputBufQueue_.empty() || bufIndexQueue_.empty()) {
                 continue;
             }
-            bufferIndex = bufIndexQueue_.front();
+            bufferIndex = (int32_t)bufIndexQueue_.front();
             bufIndexQueue_.pop();
             audioData = inputBufQueue_.front();
             inputBufQueue_.pop();
